Feat: endpoint para listar cursos com categoria

+++ techPoints/techPoints/src/main/kotlin/techForAll/techPoints/dtos/SessoesDto.kt
@@ -1,12 +1,12 @@
 package techForAll.techPoints.dtos
 
 
-class SessoesDto (
+class SessoesDto(
     var id: Long,
 
     var diaSessao: String,
 
-    var qtdTempoSessao: Double,
+    var qtdTempoSessao: Int,
 
     var metaEstudoSemana: Long
 )
